同一size type到uint32，节省一些空间，并减少了一些转换开销

+++ s_flat_hash_map/bench/bench_default_vs_no_check_dup.cpp
@@ -155,7 +155,7 @@ inline void bench_emplace(Map &map, const typename Map::key_type &key, std::uint
 }
 
 template <bool UseEmplaceNew, class Map, class Key> void fill_map_with_keys(Map &map, const std::vector<Key> &keys) {
-    map.reserve(keys.size());
+    map.reserve(static_cast<UInt32>(keys.size()));
     for (std::size_t i = 0; i < keys.size(); ++i) {
         bench_emplace<UseEmplaceNew>(map, keys[i], static_cast<std::uint32_t>(i));
     }
